changes from main + homework4
